+++ Apps/Examples/Examples/SceneDelegate.swift
@@ -1,4 +1,3 @@
-import FioriSwiftUICore
 import SwiftUI
 import UIKit
